test curso

+++ angular-test/src/app/components/user/user.component.ts
@@ -8,8 +8,11 @@ import { User } from 'src/app/models/user.model';
     templateUrl:'./user.component.html'
 })
 export class UserComponent implements OnInit{
-    
+    question: string;
+    isActive: boolean;
     users: User[];
+    userToDelete: User;
+
     ngOnInit(): void {
         this.users=[];
         this.getUsers();
@@ -27,15 +30,24 @@ export class UserComponent implements OnInit{
 
     deleteUser(user: User){
         console.log("user en padre->",user);
-        //this.products = this.products.filter(prd => prd !== prod);
-        const { id } = user;
-        this.userService.deleteUser(id)
-            .subscribe(response =>{
-            console.log(response);
-            //this.products = this.products.filter(prd => prd.id !== prod.id);
-            this.getUsers();
-        });
+        
+        this.question = `¿Esta seguro que desea eliminar el usuario <b>"${user.name}"</b>? `;
+        this.isActive = true;
+        this.userToDelete = user;
+    }
 
+    onResponse(confirmation: boolean) {
+
+        if (confirmation) {
+            const { id } = this.userToDelete;
+            this.userService.deleteUser(id)
+                .subscribe(response =>{
+                console.log(response);
+                this.getUsers();
+            });
+        } else {
+            this.isActive = false;
+        }
     }
 
 }
